+++ .dockerignore
@@ -1,8 +1,3 @@
-Dockerfile
-README.md
-LICENSE
-.dockerignore
-.env
-.idea
-__pycache__
-.git
+*
+!/app
+!requirements.txt
