+++ app/routers/moviesRoute.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from controllers.moviesController import movie_router
+from app.controllers.moviesController import movie_router
 
 routerMovie = APIRouter()
 routerMovie.include_router(movie_router)
